LayoutTests/fast/forms/auto-fill-button/show-auto-fill-button-while-selection-inside-field-expected.txt

This test ensures clicking on an input field, and then adding an autofill button, leaves the selection inside the field.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.value is "Test"
PASS successfullyParsed is true

TEST COMPLETE
